+++ src/stb_image.h
@@ -1864,7 +1864,7 @@ static float   *stbi__ldr_to_hdr(stbi_uc *data, int x, int y, int comp)
    if (comp & 1) n = comp; else n = comp-1;
    for (i=0; i < x*y; ++i) {
       for (k=0; k < n; ++k) {
-         output[i*comp + k] = (float) (pow(data[i*comp+k]/255.0f, stbi__l2h_gamma) * stbi__l2h_scale);
+         output[i*comp + k] = powf(data[i*comp+k]/255.0f, stbi__l2h_gamma) * stbi__l2h_scale;
       }
    }
    if (n < comp) {
